separate ViewBinding logic out of BaseActivity into a BaseBindingActivity
this will allow us to still use other BaseActivity functionality with Compose as we move away from view & data binding
