+++ ui/base/src/test/kotlin/org/cru/godtools/base/ui/activity/BaseActivityTest.kt
@@ -2,20 +2,16 @@ package org.cru.godtools.base.ui.activity
 
 import android.app.Application
 import android.os.Handler
-import android.view.View
 import androidx.appcompat.R
 import androidx.lifecycle.Lifecycle
 import androidx.test.core.app.ActivityScenario
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.viewbinding.ViewBinding
 import org.ccci.gto.android.common.base.TimeConstants.DAY_IN_MS
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mockito.kotlin.doAnswer
-import org.mockito.kotlin.mock
 import org.robolectric.Robolectric
 import org.robolectric.android.controller.ActivityController
 import org.robolectric.annotation.Config
@@ -24,7 +20,7 @@ import org.robolectric.annotation.Config
 @Config(application = MockApplication::class)
 class BaseActivityTest {
     lateinit var controller: ActivityController<ConcreteBaseActivity>
-    lateinit var activity: BaseActivity<*>
+    lateinit var activity: BaseActivity
 
     @Before
     fun setup() {
@@ -56,12 +52,4 @@ class MockApplication : Application() {
     }
 }
 
-class ConcreteBaseActivity : BaseActivity<ViewBinding>() {
-    init {
-        settings = mock()
-    }
-
-    override fun inflateBinding() = mock<ViewBinding> {
-        on { root } doAnswer { View(this@ConcreteBaseActivity) }
-    }
-}
+class ConcreteBaseActivity : BaseActivity()
